move chat from types to models
